pyproject.toml
[tool.poetry]
name = "geodata"
version = "0.1.0"
description = ""
authors = ["Jon Wilberg <jon.r.wilberg@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10, <3.11"
apache-airflow = "^2.10.2"
h3 = "^4.1.1"
geopandas = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
jupyterlab = "^4.2.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
